ingress/requirements.txt
pyrebase4
tensorflow
numpy
